fix user id data type

+++ src/main/java/com/wasd/web/controller/UserController.java
@@ -24,7 +24,7 @@ public class UserController {
     }
     
     @GetMapping("{id}")
-    public UserResponse findById(@PathVariable Integer id) {
+    public UserResponse findById(@PathVariable Long id) {
         return userService.findById(id);
     }
     
@@ -34,13 +34,13 @@ public class UserController {
     }
     
     @PatchMapping("/{id}")
-    public UserResponse update(@PathVariable Integer id, @RequestBody UserRequest request) {
+    public UserResponse update(@PathVariable Long id, @RequestBody UserRequest request) {
         return userService.update(id, request);
     }
     
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public void delete(@PathVariable Long id) {
         userService.delete(id);
     }
 }
